+++ msgmover/msgmover.py
@@ -1,22 +1,21 @@
+import asyncio
+import aiohttp
+import json
+import typing
+
 from redbot.core import Config, commands, checks
 from redbot.core.utils.predicates import ReactionPredicate
 from redbot.core.utils.menus import start_adding_reactions
-from urllib.request import urlopen
-import mimetypes
 import discord
 from discord import Webhook, SyncWebhook
-import aiohttp
 import requests
 
-# default modules
-import asyncio
-import base64
-from io import BytesIO
-import json
-import random
-import textwrap
-import traceback
-import typing
+from .utils import msgFormatter, webhookSettings, webhookFinder, WEBHOOK_EMPTY_AVATAR, WEBHOOK_EMPTY_NAME
+from .utils_copy import timestampEmbed
+from .utils_relay import relayGetData, relayAddChannel, relayRemoveChannel, relayCheckInput
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 class Msgmover(commands.Cog):
@@ -55,111 +54,10 @@ class Msgmover(commands.Cog):
         pass
 
 
-    # Utility Commands
-
-    def relayGetData(self, json):
-        relayInfo = {
-            "toWebhook": json.get("toWebhook", ""),
-            "attachsAsUrl": json.get("attachsAsUrl", True),
-            "userProfiles": json.get("userProfiles", True),
-        }
-        return relayInfo
-
-    async def relayAddChannel(self, ctx, chanObj, toWebhook):
-        msgrelayStoreV2 = await self.config.guild(ctx.guild).msgrelayStoreV2()
-        
-        # Set attachsAsUrl
-        attachsAsUrl = await ctx.send("Do you want attachments (images) to be forwarded as links?\n> **Yes:** Images will be sent as links\n> **No:** Images will be re-uploaded as a new file")
-        start_adding_reactions(attachsAsUrl, ReactionPredicate.YES_OR_NO_EMOJIS)
-        attachsAsUrlPred = ReactionPredicate.yes_or_no(attachsAsUrl, ctx.author)
-        await ctx.bot.wait_for("reaction_add", check=attachsAsUrlPred)
-
-        # Set userProfiles
-        userProfiles = await ctx.send("Do you want messages to be forwarded with profiles?\n> **Yes:** Messages will be forwarded with usernames and profile pics\n> **No:** Messages will use the webhook's name and image instead")
-        start_adding_reactions(userProfiles, ReactionPredicate.YES_OR_NO_EMOJIS)
-        userProfilesPred = ReactionPredicate.yes_or_no(userProfiles, ctx.author)
-        await ctx.bot.wait_for("reaction_add", check=userProfilesPred)
-
-        # Create data store
-        try:
-            relayInfo = {
-                "toWebhook": str(toWebhook),
-                "attachsAsUrl": attachsAsUrlPred.result,
-                "userProfiles": userProfilesPred.result,
-            }
-        except:
-            return False
-
-        # Append to data
-        try:
-            msgrelayStoreV2[str(chanObj.id)].append(relayInfo)
-        except KeyError:
-            msgrelayStoreV2[str(chanObj.id)] = [relayInfo]
-        except:
-            return False
-
-        # Return changes
-        return msgrelayStoreV2
-
-    async def relayRemoveChannel(self, ctx, channel, itemToDelete):
-        # Retrieve stored data
-        msgrelayStoreV2 = await self.config.guild(ctx.guild).msgrelayStoreV2()
-
-        # Remove from data
-        if itemToDelete <= 0:
-            msgrelayStoreV2.pop(str(channel.id), None)
-        else:
-            # Zero-indexed
-            msgrelayStoreV2[str(channel.id)].pop(itemToDelete-1)
-
-        # Push changes
-        await self.config.guild(ctx.guild).msgrelayStoreV2.set(msgrelayStoreV2)
-        return True
-
-    async def relayCheckInput(self, ctx, toChannel):
-        # Find/create webhook at destination if input is a channel
-        if isinstance(toChannel, discord.TextChannel):
-            toWebhook = await self.webhookFinder(toChannel)
-            if toWebhook == False:
-                await ctx.send("An error occurred: could not create webhook. Am I missing permissions?")
-                return False
-            return toWebhook
-        # Use webhook url as-is if there is https link (doesn't have to be Discord)
-        if "https://" in toChannel:
-            return str(toChannel)
-        # Error likely occurred, return False
-        await ctx.send("Error: Channel is not in this server, or webhook URL is invalid.\n\nIf you're trying to move messages across servers, please create a Webhook in the channel you want to send to: https://support.discord.com/hc/article_attachments/1500000463501/Screen_Shot_2020-12-15_at_4.41.53_PM.png")
-        return False
-
-    async def timestampEmbed(self, ctx, utcTimeObj):
-        embedColor = await ctx.embed_colour()
-        embed = discord.Embed(color=embedColor, timestamp=utcTimeObj)
-        embed.set_footer(text='\u200b')
-        return embed
-
-    async def webhookFinder(self, channel):
-        # Find a webhook that the bot made
-        try:
-            whooklist = await channel.webhooks()
-        except:
-            return False
-        # Return if match
-        for wh in whooklist:
-            if self.bot.user == wh.user:
-                return wh.url
-        # If the function got here, it means there isn't one that the bot made
-        try:
-            newHook = await channel.create_webhook(name="Webhook")
-            return newHook.url
-        # Could not create webhook, return False
-        except:
-            return False
-
-
     # Bot Commands
 
     @commands.group()
-    @checks.guildowner_or_permissions(manage_messages=True)
+    @commands.has_permissions(manage_messages=True)
     async def msgmover(self, ctx: commands.Context):
         """Hi! Thanks for installing msgmover!
 
@@ -176,8 +74,10 @@ class Msgmover(commands.Cog):
         if not ctx.invoked_subcommand:
             pass
 
+
     @commands.command(aliases=["msgmove"])
-    @checks.guildowner_or_permissions(manage_messages=True)
+    @commands.has_permissions(manage_messages=True)
+    @commands.bot_has_permissions(add_reactions=True, read_message_history=True)
     async def msgcopy(self, ctx, fromChannel: discord.TextChannel, toChannel: typing.Union[discord.TextChannel, str], maxMessages:int, skipMessages:int=0):
         """Copies messages from one channel to another
 
@@ -187,15 +87,18 @@ class Msgmover(commands.Cog):
         
         Retrieving more than 10 messages will result in Discord ratelimit throttling, so please be patient.
         
-        *Errors? Please [help us by reporting them in our Support Discord >](https://coffeebank.github.io/discord)*"""
+        *Errors? Please [help us by reporting them in our Support Discord >](https://coffeebank.github.io/discord)*
+        """
 
         # Error catching
-        toWebhook = await self.relayCheckInput(ctx, toChannel)
+        toWebhook = await relayCheckInput(self, ctx, toChannel)
         if toWebhook == False:
             return
 
+        if maxMessages <= 0:
+            return await ctx.send("Error: Please input a valid number of messages to copy.")
         if skipMessages >= maxMessages:
-            return await ctx.send("Cannot skip more messages than the max number of messages you are retrieving.")
+            return await ctx.send("Error: Cannot skip more messages than the max number of messages you are retrieving.")
 
         # Start webhook session
         await ctx.message.add_reaction("⏳")
@@ -216,19 +119,19 @@ class Msgmover(commands.Cog):
             # If they equal, it means it's the first item, so send timestamp
             if msgItemLast == msgItem.created_at or (msgItem.created_at-msgItemLast).total_seconds() > 600:
                 webhook.send(
-                    username="\u2e33\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2e33",
-                    avatar_url='https://upload.wikimedia.org/wikipedia/commons/thumb/2/25/0-Background.svg/300px-0-Background.svg.png',
-                    embed=await self.timestampEmbed(ctx, msgItem.created_at)
+                    username=WEBHOOK_EMPTY_NAME,
+                    avatar_url=WEBHOOK_EMPTY_AVATAR,
+                    embed=await timestampEmbed(self, ctx, msgItem.created_at)
                 )
                 await asyncio.sleep(1)
-            configJson = self.relayGetData({"attachsAsUrl": False, "userProfiles": True})
-            whMsg = await self.msgFormatter(webhook, msgItem, configJson)
+            configJson = webhookSettings({"attachsAsUrl": False, "userProfiles": True})
+            whMsg = await msgFormatter(self, webhook, msgItem, configJson)
             if whMsg == False:
                 await ctx.send("Failed to send: "+str(msgItem))
             else:
                 # Trigger edited tag if it was edited
                 if msgItem.edited_at:
-                    await self.msgFormatter(webhook, msgItem, configJson, editMsgId=whMsg.id)
+                    await msgFormatter(self, webhook, msgItem, configJson, editMsgId=whMsg.id)
             # Save timestamp to msgItemLast
             msgItemLast = msgItem.created_at
 
@@ -238,12 +141,16 @@ class Msgmover(commands.Cog):
         except discord.NotFound:
             await ctx.send("Done!")
 
-    @commands.group()
-    @checks.guildowner_or_permissions(administrator=True)
-    async def msgrelay(self, ctx: commands.Context):
-        """Set message relays
 
-        Message relays allow you to forward messages to another server via a webhook.
+    @commands.group()
+    @commands.has_permissions(administrator=True)
+    @commands.bot_has_permissions(add_reactions=True, embed_links=True)
+    async def msgrelay(self, ctx: commands.Context):
+        """Create message relays
+
+        Forward new messages to another server via a webhook - as if you created a portal between the two chats.
+        
+        (Also works with Discord-webhook-compatible bridges for other chat apps.)
 
         *[Join the Support Discord for announcements and more info](https://coffeebank.github.io/discord)*"""
         if not ctx.invoked_subcommand:
@@ -266,12 +173,12 @@ class Msgmover(commands.Cog):
         Cross-server relays must be a webhook. [How to create webhooks >](https://support.discord.com/hc/article_attachments/1500000463501/Screen_Shot_2020-12-15_at_4.41.53_PM.png)"""
 
         # Error catching
-        relayCheckInput = await self.relayCheckInput(ctx, toChannel)
-        if relayCheckInput == False:
+        relayResp = await relayCheckInput(self, ctx, toChannel)
+        if relayResp == False:
             return
 
         # Create entry
-        relayAdd = await self.relayAddChannel(ctx, fromChannel, relayCheckInput)
+        relayAdd = await relayAddChannel(self, ctx, fromChannel, relayResp)
         if relayAdd == False:
             return await ctx.send("Setup was stopped. Exited.")
         else:
@@ -291,19 +198,19 @@ class Msgmover(commands.Cog):
         Currently only supports webhook urls. [How to create webhooks.](https://support.discord.com/hc/article_attachments/1500000463501/Screen_Shot_2020-12-15_at_4.41.53_PM.png)
         *[Help us develop support for #channels >](https://coffeebank.github.io/discord)*"""
         # Error catching
-        relayCheckInput = await self.relayCheckInput(ctx, toChannel)
-        if relayCheckInput == False:
+        relayResp = await relayCheckInput(self, ctx, toChannel)
+        if relayResp == False:
             return
 
         # Ask for input and have it successfully complete before deleting old entry
-        relayAdd = await self.relayAddChannel(ctx, fromChannel, relayCheckInput)
+        relayAdd = await relayAddChannel(self, ctx, fromChannel, relayResp)
         if relayAdd == False:
             return await ctx.send("Setup was stopped. Exited.")
         else:
             # Create new entry
             await self.config.guild(ctx.guild).msgrelayStoreV2.set(relayAdd)
             # Delete old entry
-            await self.relayRemoveChannel(ctx, fromChannel, itemToEdit)
+            await relayRemoveChannel(self, ctx, fromChannel, itemToEdit)
 
         # Test
         try:
@@ -320,7 +227,7 @@ class Msgmover(commands.Cog):
         itemToDelete: put 1, 2, ... for which one you want to delete.
         
         To delete all relays for a fromChannel, set itemToDelete to 0 (zero)."""
-        result = await self.relayRemoveChannel(ctx, fromChannel, itemToDelete)
+        result = await relayRemoveChannel(self, ctx, fromChannel, itemToDelete)
         if result == False:
             return await ctx.send("Deletion failed. Please report this error to the support server at <https://coffeebank.github.io/discord>.")
         await ctx.message.add_reaction("✅")
@@ -370,9 +277,9 @@ class Msgmover(commands.Cog):
                     hookData = await self.fixMsgrelayStoreV2alpha(message)
                 # Send along webhook for each in array
                 for wh in hookData:
-                    configJson = self.relayGetData(wh)
+                    configJson = relayGetData(wh)
                     webhook = SyncWebhook.from_url(wh["toWebhook"])
-                    whResult = await self.msgFormatter(webhook, message, configJson)
+                    whResult = await msgFormatter(self, webhook, message, configJson)
                     wh["whResult"] = whResult.id
                 # Wait, then check for edits/deletes
                 if relayTimer <= 0:
@@ -383,223 +290,19 @@ class Msgmover(commands.Cog):
                         endMsg = await message.channel.fetch_message(message.id)
                     except discord.NotFound:
                         for wf in hookData:
-                            configJson = self.relayGetData(wh)
+                            configJson = relayGetData(wh)
                             webhook = SyncWebhook.from_url(wf["toWebhook"])
-                            await self.msgFormatter(webhook, message, configJson, deleteMsgId=wf.get("whResult", None))
+                            await msgFormatter(self, webhook, message, configJson, deleteMsgId=wf.get("whResult", None))
                     else:
                         if endMsg.edited_at:
                             for wf in hookData:
-                                configJson = self.relayGetData(wh)
+                                configJson = relayGetData(wh)
                                 webhook = SyncWebhook.from_url(wf["toWebhook"])
-                                await self.msgFormatter(webhook, endMsg, configJson, editMsgId=wf.get("whResult", None))
+                                await msgFormatter(self, webhook, endMsg, configJson, editMsgId=wf.get("whResult", None))
             else:
                 return
         else:
             return
-
-    async def msgFormatter(self, webhook, message, json, editMsgId=None, deleteMsgId=None):
-        # webhook: A webhook object from discord.py
-        # message: A message object from discord.py
-        # json: A {dict} with config variables
-
-        # Delete the message if it's delete
-        if deleteMsgId is not None:
-            try:
-                return webhook.delete_message(deleteMsgId)
-            except:
-                return False
-
-        # Edit the message if it's edit
-        if editMsgId is not None:
-            try:
-                return webhook.edit_message(
-                    message_id=editMsgId,
-                    content=message.clean_content
-                )
-            except discord.HTTPException:
-                return webhook.edit_message(
-                    message_id=editMsgId,
-                    content="**Discord:** Unsupported content\n"+str(message.clean_content)
-                )
-            except:
-                return False
-
-        # Check for system messages, Set up user profile
-        userProfilesName = None
-        userProfilesAvatar = None
-        if message.type == discord.MessageType.default or message.type == discord.MessageType.reply:
-            msgContent = message.clean_content
-            if json["userProfiles"] == True:
-                # (dpy-v2) "Discord" is not allowed in webhook usernames
-                userProfilesName = message.author.display_name.replace("Discord", "D🗪cord").replace("discord", "d🗪cord")
-                userProfilesAvatar = message.author.display_avatar.url
-        else:
-            msgContent = "**Discord:** "+str(message.type)
-            if json["userProfiles"] == True:
-                userProfilesName = "\u2e33\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2002\u2e33"
-                userProfilesAvatar = 'https://upload.wikimedia.org/wikipedia/commons/thumb/2/25/0-Background.svg/300px-0-Background.svg.png'
-
-        # Add reply if exists
-        if message.reference and message.type == discord.MessageType.reply:
-            # Retrieve replied-to message
-            refObj = message.reference.resolved
-            replyEmbed = discord.Embed(color=discord.Color(value=0x25c059), description="")
-
-            # Fallback
-            if hasattr(refObj, "clean_content") is False:
-                refObj = message
-                refContent = "Message not found"
-                refUrl = ""
-            else:
-                refContent = refObj.clean_content
-                refUrl = refObj.jump_url
-            
-            # Fill content
-            if refContent:
-                replyBody = (refContent[:56] + '...') if len(refContent) > 56 else refContent
-            else:
-                replyBody = "Click to see attachment 🖼️"
-            # Create link to message
-            replyTitle = f"↪️ {replyBody}"
-            if json["userProfiles"] == True:
-                replyEmbed.set_author(name=replyTitle, icon_url=refObj.author.display_avatar.url, url=refUrl)
-            else:
-                replyEmbed.set_author(name=replyTitle, url=refUrl)
-            # Send this before the original message so that the embed appears above the message in chat
-            webhook.send(
-                username=userProfilesName,
-                avatar_url=userProfilesAvatar,
-                embed=replyEmbed
-            )
-            await asyncio.sleep(1)
-              
-
-        # Add embed if exists
-        msgEmbed = None
-        # Do not set the embed if it came from a http link in the message (fixes repo issue #4)
-        if message.embeds and "http" not in msgContent:
-            msgEmbed = message.embeds
-
-        # Add attachment if exists
-        msgAttach = None
-        if message.attachments:
-            attachMaxSize = 10000000 # 10MB, prev. 25MB, prev. 8MB
-            attachSuccess = False
-            if json["attachsAsUrl"] == False:
-                try:
-                    # attachMaxSize upload limit
-                    totalSize = 0
-                    for mm in message.attachments:
-                        totalSize += mm.size
-                    assert totalSize < attachMaxSize
-                except AssertionError:
-                    # See if each file is under attachMaxSize, maybe we can send individually
-                    for mm in message.attachments:
-                        try:
-                            assert mm.size < attachMaxSize
-                            webhook.send(
-                                username=userProfilesName,
-                                avatar_url=userProfilesAvatar,
-                                files=[await mm.to_file()],
-                                wait=True
-                            )
-                        except AssertionError:
-                            webhook.send(
-                                content="**Discord:** File too large\n"+str(mm.url),
-                                username=userProfilesName,
-                                avatar_url=userProfilesAvatar,
-                                wait=True
-                            )
-                    attachSuccess = True
-                else:
-                    msgAttach = [await msgA.to_file() for msgA in message.attachments]
-                    attachSuccess = True
-            # Fallback if uploads don't work
-            if json["attachsAsUrl"] == True or attachSuccess == False:
-                for msgB in message.attachments:
-                    msgContent += "\n"+str(msgB.url)
-
-        # Add activity if exists
-        if message.activity:
-            msgContent += f"\n**Discord:** Activity\n"+str(message.activity)
-        if message.application:
-            msgContent += f"\n**Discord:** {message.application.name}\n{message.application.description}"
-
-        # Add sticker if exists
-        if message.stickers:
-            for msgSticker in message.stickers:
-                if msgSticker.url is not None:
-                    msgContent += "\n"+str(msgSticker.url)
-                else:
-                    msgContent += "\n**Discord:** Sticker\n"+str(msgSticker.name)+", "+str(msgSticker.id)
-
-        # Send core message
-        whMsg = False
-
-        # (dpy-v2) Switch to argument unpacking, since passing None doesn't work anymore
-        whMsgArgs = {}
-
-        try:
-            whMsgArgs = {
-              "content": msgContent,
-              "username": userProfilesName,
-              "avatar_url": userProfilesAvatar,
-              "embeds": msgEmbed,
-              "files": msgAttach,
-              "wait": True
-            }
-            whMsg = webhook.send(
-                **{k: v for k, v in whMsgArgs.items() if v is not None}
-            )
-        except discord.HTTPException:
-            # catch HTTPException: 400 Bad Request (error code: 50035): Invalid Form Body
-            #     In content: Must be 2000 or fewer in length.
-            if len(msgContent) > 1964:
-                msgLines = textwrap.wrap(msgContent, 2000, break_long_words=True)
-                for msgLineItem in msgLines:
-                    whMsgArgs = {
-                      "content": str(msgLineItem),
-                      "username": userProfilesName,
-                      "avatar_url": userProfilesAvatar,
-                      "embeds": msgEmbed,
-                      "files": msgAttach,
-                      "wait": True
-                    }
-                    whMsg = webhook.send(
-                        **{k: v for k, v in whMsgArgs.items() if v is not None}
-                    )
-            # catch HTTPException: 400 Bad Request (error code: 50006): Cannot send an empty message
-            else:
-                try:
-                    whMsgArgs = {
-                      "content": "**Discord:** Unsupported content\n" + str(msgContent[:1964]) + (str(msgContent[1964:]) and '…'),
-                      "username": userProfilesName,
-                      "avatar_url": userProfilesAvatar,
-                      "embeds": msgEmbed,
-                      "files": msgAttach,
-                      "wait": True
-                    }
-                    whMsg = webhook.send(
-                        **{k: v for k, v in whMsgArgs.items() if v is not None}
-                    )
-                # One last try, without username or avatar
-                except:
-                    whMsgArgs = {
-                      "content": "**Discord:** Unsupported content\n" + str(msgContent[:1964]) + (str(msgContent[1964:]) and '…'),
-                      "username": "Unknown User",
-                      "embeds": msgEmbed,
-                      "files": msgAttach,
-                      "wait": True
-                    }
-                    whMsg = webhook.send(
-                        **{k: v for k, v in whMsgArgs.items() if v is not None}
-                    )
-        except:
-            traceback.print_exc()
-            return False
-        
-        # Need to tell endpoint that function ended, so that sent message order is enforceable by await
-        return whMsg
 
 
     # Legacy Commands
